backend/requirements.txt
Flask==2.3.3
celery==5.3.4
redis==5.0.1
Pillow==10.2.0
flask_cors==5.0.1
